Make Frozen type family again
Purpose of Frozen is to provide immutable snapshot of generator state from which
generator could be restored. As such it shouldn't carry state token, thus it
couldn't be injective and could only be implemented as a type family.

Current design makes it impossible to implement following patterns generically:

> runST $ do do_something_with g
>            freezeGen g

> frozen <- readGeneratorState
> return $ runST $ do g <- thawGen
>                     do_something_with g

This change does make type inference worse. But this is acceptable price to pay
in order to have useful Frozen.