Add function for generating Float/Double in [0,1] range (#149)
There are functions for Float and Double. They come in two variants: one that
sample in [0,1] interval and one that samples in (0,1].  Former allows to save
few floating point operations when one needs values in [0,1] range. Which is very
commonly used primitive

Latter is important when output of generator is fed into function which not
defined at 0. This is again very common. For Float sampling zero is
feasible. For Double it's improbable but could in principle happen in bad
generator/seed combo, or for malicious generator initialization.